updating down script
Johan, When used this up and down the script in Master branch, it was working on Lan properly and DNS issue was fixed, if you can modify it little to work on develop branch, it will be very helpful.

What I did: We are finding around to get the problem solved easily. We found that you took that code from sket29's repo instead of tunnelblick's official repo. So we took the latest code from official repo since sket29's pull request was pending for 8 months and I contacted him, he said he lost interested on his code and moved to IPSEC. So instead of making his code as based, we are now taking the official code as base. We have tested a lot in master branch, it worked. I changed string here as you changed the string of sket29's code